split handlers, added the possibility to visit personal_cab of any user on homepage and inside theme (by clicking on user's name
