+++ template/home.page.html
@@ -2,8 +2,9 @@
 
 {{define "centralPart"}}
 {{$loggedAs := index .Data "loggedAs"}}
+{{$loggedAsID := index .Data "loggedAsID"}}
 <br>
- <h3> Logged As: <a href="/personal_cabinet"> {{$loggedAs}} - go to persinal cabinet.</a></h3>
+ <h3> Logged As: <a href="/personal_cabinet?userID={{$loggedAsID}}"> {{$loggedAs}}.</a></h3>
 <div class="container custom-container">
     <table class="table border-success forumTable whiteLink mt-3">
         <thead class="containerHead">
