+++ template/personal.page.html
@@ -1,25 +1,55 @@
 {{template "body" .}}
 
 {{define "centralPart"}}
-    <h1>User Profile</h1>
-
     <div class="content">
         {{$person := index .Data "personal"}}
-        <h2>Personal Information</h2>
+        <h2>User Profile</h2>
         <form action="/update-profile" method="POST">
-            <img src="{{$person.Picture}}" alt="your_ava" class="ava_img">
-            <br>
-            <label for="username">Username: {{$person.UserName}}</label>
-            <br>
-            <label for="email">Email: {{$person.Email}}</label>
-            <br>
-            <label for="email">First Name: {{$person.FirstName}}</label>
-            <br>
-            <label for="email">Last Name: {{$person.LastName}}</label>
-            <br>
-            <label for="password">Created: {{$person.Created}}</label>
-            <br>
-            <input type="submit" value="Update">
+            <div class="container ">
+                <table class="table table-borderless whiteLink">
+                    <tbody>
+                        <tr>
+                            <td colspan="2" class="text-center">
+                                <img src="{{$person.Picture}}" alt="your_ava" class="ava_img">
+                            </td>
+                        </tr>
+                        <tr>
+                            <th>Username</th>
+                            <td>{{$person.UserName}}</td>
+                        </tr>
+                        <tr>
+                            <th>Email</th>
+                            <td>{{$person.Email}}</td>
+                        </tr>
+                        <tr>
+                            <th>First Name</th>
+                            <td>{{$person.FirstName}}</td>
+                        </tr>
+                        <tr>
+                            <th>Last Name</th>
+                            <td>{{$person.LastName}}</td>
+                        </tr>
+                        <tr>
+                            <th>Created</th>
+                            <td>{{$person.Created}}</td>
+                        </tr>
+                        <tr>
+                            <th>Total messages:</th>
+                            {{$totalP := index .Data "totalPosts"}}
+                            <td>{{$totalP}}</td>
+                        </tr>
+                        <tr>
+                            <td colspan="2" class="text-center">
+                                <form action="/update-profile" method="POST">
+                                    <input type="submit" class="btn btn-primary" value="Update">
+                                </form>
+                            </td>
+                        </tr>
+                    </tbody>
+                </table>
+            </div>
+            
         </form>
+    </div>
     </div>
 {{end}}
